+++ src/collective/jsonmigrator/blueprints/mimetype.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
+from collective.jsonmigrator.blueprints.utils import remove_first_bar
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import defaultKeys
 from collective.transmogrifier.utils import Matcher
 from collective.transmogrifier.utils import traverse
-from Products.CMFPlone.utils import safe_unicode
 from zope.interface import implementer
 from zope.interface import provider
 
@@ -47,7 +47,7 @@ class Mimetype(object):
                 yield item
                 continue
 
-            path = safe_unicode(item[pathkey].lstrip("/")).encode("ascii")
+            path = remove_first_bar(item[pathkey])
             obj = traverse(self.context, path, None)
 
             if obj is None: