+++ src/collective/jsonmigrator/blueprints/workflowhistory.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from collective.jsonmigrator.blueprints.utils import remove_first_bar
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import defaultKeys
@@ -6,7 +7,6 @@ from collective.transmogrifier.utils import Matcher
 from collective.transmogrifier.utils import traverse
 from DateTime import DateTime
 from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import safe_unicode
 from zope.interface import implementer
 from zope.interface import provider
 
@@ -61,7 +61,7 @@ class WorkflowHistory(object):
                 continue
 
             # traverse() available in version 1.5+ of collective.transmogrifier
-            path = safe_unicode(item[pathkey].lstrip("/")).encode("ascii")
+            path = remove_first_bar(item[pathkey])
             obj = traverse(self.context, path, None)
 
             if obj is None or not getattr(obj, "workflow_history", False):